Added necessary wiring to get as far as asking an arm of the analyser to check out an Acorn disk image.
