+++ StaticAnalyser/Acorn/Disk.cpp
@@ -1,0 +1,17 @@
+//
+//  Disk.cpp
+//  Clock Signal
+//
+//  Created by Thomas Harte on 18/09/2016.
+//  Copyright © 2016 Thomas Harte. All rights reserved.
+//
+
+#include "Disk.hpp"
+
+using namespace StaticAnalyser::Acorn;
+
+std::list<File> StaticAnalyser::Acorn::GetDFSFiles(const std::shared_ptr<Storage::Disk::Disk> &disk)
+{
+	std::list<File> files;
+	return files;
+}
