+++ StaticAnalyser/Acorn/File.cpp
@@ -1,0 +1,9 @@
+//
+//  File.cpp
+//  Clock Signal
+//
+//  Created by Thomas Harte on 18/09/2016.
+//  Copyright © 2016 Thomas Harte. All rights reserved.
+//
+
+#include "File.hpp"
